Change to_d to to_deg
Ruby 1.9 introduced  to_d, for to double, which conflicted with my use
of to_d as and alias for to degrees.  Hence to_d has been changed to
to_deg.

Also added to_rad as an alternate alias to to_r. The aim is to
eventually remove the to_r.